add BuildManager class to more explicitly organise the build steps
Right now, this is just a WIP because for some reason despite using the
same code, it seems to be pulling in the entire upstream repo (including
all of its papers) rather than just the proceedings build tools. I don't
know why it works for the standard build script but not the main build
script.