BUG: stop keying papers with enumerate
procbuild assigns a unique identifier to each paper that it uses
to retrieve the paper status and pdf. Currently, this is done
by enumerating open PRs, which works fine until a PR is opened
or closed, at which point all of the IDs are off by one. This
commit replaces the integer keys with one composed of the user
and branch name, which is hopefully close enough to unique.
Closes #18.

Note: the commit is not as large as the diff seems to indicate.
A lot of the file changes are from my editor auto-deleting the
trailing whitespace on a bunch of lines.