BUG: clone repo using real repo name
In the user PR code cloning logic, we currently assume
that the repo name is always scipy_proceedings. This causes
procbuild to fail when a user forks our repo and changes the
name, e.g. to scipy_proceedings_2023. This changes the PR
listing logic to fetch and return the actual repo name, and
then passes that through the BuildManager into the clone
url function.

xref https://github.com/scipy-conference/scipy_proceedings/pull/831